// src/model/car/Info/ColorOption.ts
export interface ColorOption {
  colorOptionPk: string;
  carPk: string;
  carCountryName: string;
  carBrandName: string;
  carModelName: string;
  carDetailModelName: string;
  colorOptionType: string;
  colorOptionPrice: number;
  colorOptionName: string;
  colorOptionCodes: string[];
  colorOptionCodeCount: number;
  removeExteriorColors: [];
  removeInteriorColors: [];
}